+++ site_files/app/Http/Controllers/Back/StateController.php
@@ -52,17 +52,27 @@ class StateController extends Controller
                 return $str;
             })
             ->addColumn('action', function ($states) {
+                $editStr = '';
+                $deleteStr = '';
+
+                if (isAllowed('Can Edit State')) {
+                    $editStr = '<li>
+                    <a href="' . route('states.edit', ['stateObj' => $states->id]) . '" class="text-info"><i class="fas fa-edit" aria-hidden="true"></i>Edit</a>
+                </li>';
+                }
+                if (isAllowed('Can Delete State')) {
+                    $deleteStr = '<li>
+                    <a href="javascript:void(0);" onclick="deleteState(' . $states->id . ');" class="text-danger"><i class="fas fa-trash" aria-hidden="true"></i>Delete</a>
+                </li>';
+                }
+
                 return '
                 <div class="btn-group">
 					<button class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">Action
 					</button>
 					<ul class="dropdown-menu">
-						<li>
-							<a href="' . route('states.edit', ['stateObj' => $states->id]) . '" class="text-info"><i class="fas fa-edit" aria-hidden="true"></i>Edit</a>
-						</li>
-						<li>
-							<a href="javascript:void(0);" onclick="deleteState(' . $states->id . ');" class="text-danger"><i class="fas fa-trash" aria-hidden="true"></i>Delete</a>
-						</li>
+						' . $editStr . '
+						' . $deleteStr . '
 					</ul>
 				</div>';
             })
@@ -83,7 +93,7 @@ class StateController extends Controller
      */
     public function create()
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Add State');
         $title = FindInsettingArr('business_name') . ': States Management';
         $msg = '';
         $stateObj = new State();
@@ -101,7 +111,7 @@ class StateController extends Controller
      */
     public function store(StateBackFormRequest $request)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Add State');
         $stateObj = new State();
         $stateObj = $this->setStateValues($request, $stateObj);
         $stateObj->save();
@@ -129,7 +139,7 @@ class StateController extends Controller
      */
     public function edit(State $stateObj)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Edit State');
         $title = FindInsettingArr('business_name') . ': States Management';
         $msg = '';
         return view('back.states.edit')
@@ -147,7 +157,7 @@ class StateController extends Controller
      */
     public function update(StateBackFormRequest $request, State $stateObj)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Edit State');
         $stateObj = $this->setStateValues($request, $stateObj);
         $stateObj->save();
 
@@ -165,7 +175,7 @@ class StateController extends Controller
      */
     public function destroy(State $stateObj)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Delete State');
         County::where('state_id', $stateObj->id)->delete();
         City::where('state_id', $stateObj->id)->delete();
         $stateObj->delete();
@@ -174,7 +184,7 @@ class StateController extends Controller
 
     public function updateStateStatus(Request $request)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Edit State');
         $stateObj = State::find($request->id);
         $stateObj = $this->setStateStatus($request, $stateObj);
         $stateObj->update();
@@ -183,7 +193,7 @@ class StateController extends Controller
 
     public function sortStates()
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Sort States');
         $title = FindInsettingArr('business_name') . ': States Management';
         $msg = '';
         return view('back.states.sort')->with('title', $title)
@@ -192,7 +202,7 @@ class StateController extends Controller
 
     public function statesSortData(Request $request)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Sort States');
         $states = State::select('states.id', 'states.state_name', 'states.sort_order')
             ->where('status', 'like', 'active')
             ->orderBy('sort_order', 'ASC')->get();
@@ -207,7 +217,7 @@ class StateController extends Controller
 
     public function statesSortUpdate(Request $request)
     {
-        hasPermission('Can Manage States');
+        hasPermission('Can Sort States');
         $statesOrder = $request->input('statesOrder');
         $statesOrderArray = explode(',', $statesOrder);
         $count = 1;
